+++ CHANGELOG.md
@@ -207,6 +207,7 @@ This change was introduced in [pull request #1801: Add Nunjucks options for comp
 - [#1797: Ensure components pass label.id down to the label](https://github.com/nhsuk/nhsuk-frontend/pull/1797)
 - [#1807: Fix summary list cells becoming vertically misaligned when a multi-line inline-block element is present](https://github.com/nhsuk/nhsuk-frontend/pull/1807)
 - [#1829: Always validate component config schema](https://github.com/nhsuk/nhsuk-frontend/pull/1829)
+- [#1854: Fix for nested lists in the contents list](https://github.com/nhsuk/nhsuk-frontend/pull/1854)
 
 ## 10.3.1 - 19 January 2026